server/pyproject.toml
[tool.poetry]
name = "hako"
version = "0.1.0"
description = ""
authors = ["James Chuong <chuong.james@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-ninja = "^0.21.0"
ruff = "^0.0.262"
black = "^23.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py311']
extend-exclude = '''
	/migrations/
'''
